Change all "#!/usr/bin/env python -tt" to "#!/usr/bin/python -tt".
What I'd really like to do is come up with a platform-independent way of
specifying the Python compiler at the tops of each file.  The env trick
doesn't work when passing -tt, but even if it did it's not necessarily
portable because env may not be in /usr/bin.  Also, there's no support
in these #! lines for non-Unix systems.